refactor: separate hooks, reduce logic in component
